+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 Version 1.4-SNAPSHOT
 -------------
 
+ADDED:
+
+- new factory class created for generate ParameterizedTypeReference for collection types
+- new class for create url for the rest template
+
 Version 1.3
 -------------